Significant change to Fix command - better handling of DOFs. Type hinting for list of objects.

+++ o3seespy/command/test.py
@@ -14,7 +14,7 @@ class NormUnbalance(TestBase):
     """
     op_type = 'NormUnbalance'
 
-    def __init__(self, osi, tol, max_iter, p_flag=0, n_type=2, max_incr: int=None):
+    def __init__(self, osi, tol, max_iter, p_flag=0, n_type=2, max_incr: float=None):
         """
         Initial method for NormUnbalance
